mysql backup

+++ spedia/src/main/java/com/spedia/dao/UserDao.java
@@ -3,7 +3,7 @@ package com.spedia.dao;
 import com.spedia.model.User;
 
 
-public interface UserDao extends BaseDao<User, Long>  {
+public interface UserDao extends BaseDao<User, Integer>  {
 
 	User findByUserName(String username);
